refactor: toss dto 패키지 구조 정리 (#83)
